+++ app/renderer-process/executor.js
@@ -283,29 +283,13 @@ function execTable2Cfg(params) {
     --outfile "${params.outfile}" 1> "${params.logdir}/table2cfg.log" `;
     execSyncProcess('creating config files', cmd);    
 };
-function validSnakeMake(params) {
-    let smkfile = `${process.env.ISANXOT_SRC_HOME}/wfs/wf_sanxot.smk`;
-    let cmd_smk = `"${process.env.ISANXOT_LIB_HOME}/python/tools/Scripts/snakemake.exe" \
-    --configfile "${params.configfile}" \
-    --snakefile "${smkfile}" \
-    --cores ${params.nthreads} \
-    --directory "${params.directory}"`;
-    let cmd_unlock = `${cmd_smk} --unlock `;
-    let cmd_clean  = `${cmd_smk}  --cleanup-metadata "${smkfile}"`;
-    // Sync process that Unlock the output directory
-    // First we unlock the ouput directory
-    let cmd = `${cmd_unlock} && ${cmd_clean}`
-    // let cmd = `${cmd_unlock}`
-    execSyncProcess('preparing the workspace', cmd);
-}
 function execSnakeMake(params) {
     let smkfile = `${process.env.ISANXOT_SRC_HOME}/wfs/wf_sanxot.smk`;
-    let cmd_smk = `"${process.env.ISANXOT_LIB_HOME}/python/tools/Scripts/snakemake.exe" \
+    let cmd_smk = `"${process.env.ISANXOT_LIB_HOME}/python/tools/python" "${process.env.ISANXOT_SRC_HOME}/wfs/mysnake.py" \
     --configfile "${params.configfile}" \
     --snakefile "${smkfile}" \
     --cores ${params.nthreads} \
-    --directory "${params.directory}" \
-    --rerun-incomplete --keep-going`;
+    --directory "${params.directory}" `;
     let cmd = `${cmd_smk} 1> "${params.logfile}" 2>&1`;
     execProcess('executing the workflow', cmd, params.logfile, params.directory);
 }
@@ -335,10 +319,6 @@ function saveProject(wf_date_id, wf) {
     let [cfg_dir, dte_dir, log_dir] = preparePrjWorkspace(wf_date_id, outdir, wf['prj_workspace']);
     // create datafiles
     let attfile = createConfigFiles(wf_date_id, outdir, dte_dir, wf);
-    return [outdir, cfg_dir, dte_dir, log_dir, attfile];
-};
-// validate project
-function validateProject(dte_dir, attfile, log_dir, cfg_dir, outdir) {
     // Exec: create config file for the execution of workflow
     execTable2Cfg({
         'indir': dte_dir,
@@ -348,13 +328,7 @@ function validateProject(dte_dir, attfile, log_dir, cfg_dir, outdir) {
     });
     // Copy only the files that are different
     copyDiffFiles(`${dte_dir}`, `${cfg_dir}`);
-    // Validate snakemake
-    validSnakeMake({
-        'configfile': `${outdir}/.isanxot/config.yaml`,
-        'nthreads':   `${document.querySelector('#nthreads').value}`,
-        'directory':  `${outdir}`,
-        'logfile':    `${log_dir}/isanxot.log`
-    });
+    return [outdir, cfg_dir, dte_dir, log_dir, attfile];
 };
 // execute project
 function executeProject() {
@@ -365,8 +339,6 @@ function executeProject() {
         if ( wf_date_id && wf ) {
             // save project
             let [outdir, cfg_dir, dte_dir, log_dir, attfile] = saveProject(wf_date_id, wf);
-            // validate project
-            validateProject(dte_dir, attfile, log_dir, cfg_dir, outdir);
             // Execute snakemake
             execSnakeMake({
                 'configfile': `${outdir}/.isanxot/config.yaml`,
@@ -378,34 +350,8 @@ function executeProject() {
         else {
             exceptor.showMessageBox('error', "The workflow identifier is not defined", title='No project to execute', end=true);
         }
-        // // imported variables
-        // let wf = importer.wf;
-        // let wf_date_id = importer.wf_date_id;
-        // // get the output directory
-        // let outdir = $(`#main_inputs #outdir`).val();
-        // // prepare the workspace of project
-        // let [cfg_dir, dte_dir, log_dir] = preparePrjWorkspace(wf_date_id, outdir, wf['prj_workspace']);
-        // // create datafiles
-        // let attfile = createConfigFiles(wf_date_id, outdir, dte_dir, wf);
-        // // Exec: create config file for the execution of workflow
-        // execTable2Cfg({
-        //     'indir': dte_dir,
-        //     'attfile': attfile,
-        //     'outfile': `${dte_dir}/config.yaml`,
-        //     'logdir': log_dir
-        // });
-        // // Copy only the files that are different
-        // copyDiffFiles(`${dte_dir}`, `${cfg_dir}`);
-        // // Exec: execute snakemake
-        // execSnakeMake({
-        //     'configfile': `${outdir}/.isanxot/config.yaml`,
-        //     'nthreads':   `${document.querySelector('#nthreads').value}`,
-        //     'directory':  `${outdir}`,
-        //     'logfile':    `${log_dir}/isanxot.log`
-        // });
     }, 1000); // due the execSync block everything, we have to wait until loading event is finished
 };
 module.exports = {
-    saveProject: saveProject,
-    validateProject: validateProject
+    saveProject: saveProject
 };