+++ src/preSanXoT/ratios.py
@@ -121,7 +121,12 @@ def main(args):
     # ddf = calculate_ratio(list(ddf.groupby("Experiment")), ratios)
     
     logging.info('print output')
-    ddf.to_csv(args.outfile, sep="\t", index=False)
+    # print to tmp file
+    f = f"{args.outfile}.tmp"
+    ddf.to_csv(f, sep="\t", index=False)
+    # rename tmp file
+    os.rename(f, os.path.splitext(f)[0])
+